Get adjusted fvps for particular touchstone and vaccine delivery methods

Get adjusted fvps for particular touchstone and vaccine delivery methods

get_fvps(
  con,
  touchstone,
  baseline_vaccine_delivery,
  focal_vaccine_delivery,
  focal_scenario_type,
  countries = NULL,
  vaccination_years = NULL
)

Arguments

con

DB connection

baseline_vaccine_delivery

A list of lists containing vaccine and activity_type (routine or campaign) describing delivery

focal_vaccine_delivery

A list of lists containing vaccine and activity_type (routine or campaign) describing delivery

focal_scenario_type

A character determining from which scenario_type coverage estimates shall be extracted

countries

Optional vector of countries to filter data

vaccination_years

Option vector of years to filter data

touchtonse

Touchstone to get data for

Value

Tibble containing fvp data